backend/requirements.txt
# Core Framework - Production Ready
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.2

# Machine Learning & Data Science
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.2
scipy==1.11.4

# Federated Learning - Top Tier
flwr==1.11.0
torch==2.1.2
torchvision==0.16.2
tensorflow==2.16.1

# Security & Networking
cryptography==42.0.8
PyJWT==2.9.0
bcrypt==4.1.2
python-multipart==0.0.12

# System Monitoring
psutil==6.1.0
websockets==12.0
aiofiles==24.1.0

# Database & Storage
sqlalchemy==2.0.25
alembic==1.13.1
redis==5.0.1

# HTTP & API
httpx==0.26.0
requests==2.32.3
aiohttp==3.9.1

# Configuration & Environment
python-dotenv==1.0.1
pyyaml==6.0.1
click==8.1.7

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.1.1
ruff==0.1.9

# Monitoring & Observability
prometheus-client==0.19.0
structlog==23.2.0

# Additional ML Libraries
joblib==1.3.2
matplotlib==3.9.2
seaborn==0.13.2